requirements.txt
alabaster==0.7.12
asgiref==3.2.7
Babel==2.8.0
beautifulsoup4==4.9.0
bootstrap-py==1.0.1
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.6
django-bootstrap-form==3.4
django-bootstrap4==1.1.1
django-stdimage==5.1.1
docutils==0.16
gitdb==4.0.5
GitPython==3.1.2
gunicorn==20.0.4
idna==2.9
imagesize==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==20.3
pguard==0.4.0
Pillow==7.1.2
progressbar2==3.51.3
psycopg2-binary==2.8.5
Pygments==2.6.1
PyMySQL==0.9.3
pyparsing==2.4.7
python-utils==2.4.0
pytz==2020.1
requests==2.23.0
six==1.14.0
smmap==3.0.4
snowballstemmer==2.0.0
soupsieve==2.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
